% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{source_to_supplier}
\alias{source_to_supplier}
\title{Map a source_system value to a supplier code}
\usage{
source_to_supplier(source_system)
}
\arguments{
\item{source_system}{Character vector. One or more \code{source_system} values
from the gauge registry.}
}
\value{
Character vector of supplier codes, NA for unrecognised values.
}
\description{
Converts the \code{source_system} column values used in the gauge registry
to the EA-standard supplier codes used in dataset IDs.
}
\examples{
source_to_supplier("HDE")     # "EA"
source_to_supplier("WISKI")   # "WISKI"
}
